+++ src/plugin-utils.ts
@@ -1,3 +1,0 @@
-export function escapeRegExp(string: string) {
-	return new RegExp(string.replace(/[.*+?^${}()|[\]\\]/g, "\\$&"));
-}